docs and readme update
