+++ Cargo.toml
@@ -1,17 +1,17 @@
 [package]
 name = "cronframe"
-version = "0.1.2"
+version = "0.1.3"
 authors = ["Antonio Cimino"]
 edition = "2021"
-description = "A library for creating cron jobs with macros from functions and struct types, with service web server."
+description = "A Macro Annotation Cron Job Framework with Web Server and CLI Tool."
 repository = "https://github.com/antcim/cronframe"
 readme = "README.md"
-keywords = ["cron", "cronjob", "cronjobmacro"]
+keywords = ["cron", "cronjob", "cronjobmacro", "cronjobtool"]
 license = "MIT OR Apache-2.0"
 exclude = ["examples", "tests"]
 
 [dependencies]
-cronframe_macro = { path = "./cronframe_macro", version = "0.1.2" }
+cronframe_macro = { path = "./cronframe_macro", version = "0.1.3" }
 rocket = { version = "0.5.1" }
 cron = "0.12.1"
 chrono = "0.4.38"
