+++ cronframe_macro/CHANGELOG.md
@@ -1,7 +1,12 @@
+# Version 0.1.3 - Released ---
+Derivation of Clone trait in the cron_obj macro itself.
+
+<!-- version separator -->
+
 # Version 0.1.2 - Released 2024-08-05
-Macro generated code dependency fixes
+Macro generated code dependency fixes.
 
 <!-- version separator -->
 
 # Version 0.1.1 - Released 2024-07-23
-Crate dependency fixes
+Crate dependency fixes.
