+++ cronframe_macro/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cronframe_macro"
-version = "0.1.2"
+version = "0.1.3"
 edition = "2021"
 description = "The Macro used in the cronframe crate."
 repository = "https://github.com/antcim/cronframe"
